back-end/requirements.txt
black
python-dotenv
flask
flask-cors
flask-marshmallow
flask-restful
flask-restless
flask-SQLAlchemy
marshmallow-sqlalchemy
psycopg2
requests
SQLAlchemy
selenium
